Split bio models into 8 files as the bio.py became too big
