+++ website/models/bio/__init__.py
@@ -1,0 +1,6 @@
+from .model import BioModel
+from .mutations import *
+from .protein import *
+from .gene import *
+from .enzymes import *
+from .sites import *
